text-classification/code/requirements.txt
google-cloud-bigquery>=3.9.0
google-cloud-aiplatform
flask==2.3.2